requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0

# Machine Learning
tensorflow==2.18.0
tensorflow-hub==0.15.0
numpy==1.26.2
scikit-learn==1.3.2

# Image Processing
pillow==10.1.0
opencv-python==4.8.0..74
opencv-python-headless==4.8.0.74

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Data Science
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Dataset Download
kaggle==1.5.16
roboflow==1.1.9
requests==2.31.0
urllib3==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging
colorlog==6.8.0
